requirements.txt
streamlit
google-cloud-aiplatform
pdfplumber
google-generativeai
Pillow